softing: fix compiler warnings
.../kernel/2.6/drivers/net/can/softing/softing_main.c:547:
warning: format '%04lx' expects type 'long unsigned int', but argument 4
has type 'int'

.../kernel/2.6/drivers/net/can/softing/softing_fw.c:740:
 warning: integer constant is too large for 'long' type

+++ branches/netlink/kernel/2.6/drivers/net/can/softing/softing_main.c
@@ -545,7 +545,7 @@ static int boot_card(struct softing *card)
 			continue;
 		/* memory is not equal */
 		dev_alert(card->dev, "write to dpram failed at 0x%04lx\n",
-			lp - card->dpram.virt);
+			(unsigned long)(lp - card->dpram.virt));
 		goto open_failed;
 	}
 	wmb();
